+++ tests/framework/db/pgsql/QueryBuilderTest.php
@@ -359,15 +359,23 @@ class QueryBuilderTest extends \yiiunit\framework\db\QueryBuilderTest
         $qb = $this->getQueryBuilder();
 
         $expected = 'DROP INDEX "index"';
-        $sql = $qb->dropIndex('index', 'table');
+        $sql = $qb->dropIndex('index', '{{table}}');
         $this->assertEquals($expected, $sql);
 
         $expected = 'DROP INDEX "schema"."index"';
-        $sql = $qb->dropIndex('index', 'schema.table');
+        $sql = $qb->dropIndex('index', '{{schema.table}}');
         $this->assertEquals($expected, $sql);
 
         $expected = 'DROP INDEX "schema"."index"';
-        $sql = $qb->dropIndex('schema.index', 'schema2.table');
+        $sql = $qb->dropIndex('schema.index', '{{schema2.table}}');
+        $this->assertEquals($expected, $sql);
+
+        $expected = 'DROP INDEX "schema"."index"';
+        $sql = $qb->dropIndex('index', '{{schema.%table}}');
+        $this->assertEquals($expected, $sql);
+
+        $expected = 'DROP INDEX {{%schema.index}}';
+        $sql = $qb->dropIndex('index', '{{%schema.table}}');
         $this->assertEquals($expected, $sql);
     }
 }